maybetemplate pauseMenuDrawStatus: (u32)airMeter==0 -> cmplwi (96.78->97.02)
#3 unsigned-compare lever. The airMeter==0 term in the case 10-12 && guard
emitted cmpwi (signed); target reads the global via lwz + cmplwi. Casting the
single compare operand to u32 flips it to the matching cmplwi with no other
codegen change. Pure-boolean, value not reused. Unit 96.61->96.66.

(cherry picked from commit 6791d7725a04dd9c12f0f397efbe9da06282e975)

+++ src/main/dll/maybetemplate.c
@@ -603,7 +603,7 @@ void pauseMenuDrawStatus(void)
             off = i * 4;
             if (((((f32*)(base + 0xAFC))[i] >= lbl_803E1E3C) &&
                     ((*(u16*)(player + 0xB0) & 0x1000) == 0) && (pauseMenuState == 0) &&
-                    (airMeter == 0) && (getHudHiddenFrameCount() == 0) &&
+                    ((u32)airMeter == 0) && (getHudHiddenFrameCount() == 0) &&
                     (PMDS_CAMERA_GET_STATE() != 0x44)) ||
                 ((i == 3) && ((lbl_803DD792 & 2) != 0)))
             {
